refactor base crud repository
create base repository, read only repository, and base crud repository

some repositories are read only or doesnt need some of the methods in crud i.e only create etc
